test: change promise to async/await in debugger-watcher
PR-URL: https://github.com/nodejs/node/pull/44687
